+++ src/main/java/com/example/wantedpreonboarding/page/dto/BreadCrumbsDto.java
@@ -1,0 +1,6 @@
+package com.example.wantedpreonboarding.page.dto;
+
+public interface BreadCrumbsDto {
+    Long getId();
+    String getTitle();
+}
